+++ src/electron.d.ts
@@ -3,3 +3,8 @@ declare module "electron" {
     openPath(path: string): Promise<string>;
   };
 }
+
+declare module "*.live-photo-tool" {
+  const base64: string;
+  export default base64;
+}